feat(profiles): remove dispute model and update auth
